pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crypto-scanner"
version = "0.1.0"
description = "CLI tool for scanning cryptographic usage and generating quantum-vulnerability risk assessments"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Quantum Shield Labs", email = "contact@quantumshieldlabs.dev" }
]
keywords = ["cryptography", "security", "quantum", "scanner", "vulnerability"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Security :: Cryptography",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "cryptography>=41.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
crypto-scanner = "crypto_scanner.cli:app"

[project.urls]
Homepage = "https://github.com/mbennett-labs/crypto-scanner"
Repository = "https://github.com/mbennett-labs/crypto-scanner"
Documentation = "https://github.com/mbennett-labs/crypto-scanner#readme"

[tool.hatch.build.targets.wheel]
packages = ["src/crypto_scanner"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
